[Refactor] URLFormatter를 다른 객체 프로퍼티로 가지고 있는 방식에서 URLFormattable extension을 통한 기본구현메소드로 변경
