+++ WeatherForecast/WeatherForecast/ViewController.swift
@@ -7,12 +7,9 @@
 import UIKit
 
 class ViewController: UIViewController {
-
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view.
     }
-
-
 }
